Create main.py
//Inputs:
Starting node
Destination node
Traffic conditions on each edge
5
//Outputs:
Shortest path from the starting node to the destination node
Steps:
 Initialize an array, dist[], to represent the distance from the starting node
to all other nodes in the graph.  Initialize an array, prev[], to represent the previous node in the shortest
path to each node.  Set the distance of the starting node to 0 and the distances of all other
nodes to infinity.  For each node in the graph, calculate the minimum distance from the starting
node to that node using the Dynamic Programming technique.  Update the dist[] array with the minimum distances.  Update the prev[] array with the previous node in the shortest path to each
node.  Use the prev[] array to determine the shortest path from the starting node to
the